+++ sequana/scripts/main.py
@@ -31,7 +31,7 @@ from easydev.console import red, purple, green, blue
 from easydev import DevTools, SmartFormatter
 
 
-adapters_choice = ["Nextera", "Rubicon", "PCRFree"]
+adapters_choice = ["Nextera", "Rubicon", "PCRFree", "TruSeq"]
 
 help_input = """Missing input data.
 
@@ -77,7 +77,6 @@ class Tools(object):
         self.dv.mkdir(name)
     def info(self, txt, force=False):
         if self.verbose or force: print(txt)
-
 
 
 class Options(argparse.ArgumentParser):
@@ -593,6 +592,7 @@ def copy_config_from_sequana(module, source="config.yaml",
         txt = "copied %s from sequana %s pipeline"
         print(txt % (source, module.name))
 
+
 def sequana_init(options):
     import sequana
     from sequana.misc import textwrap
@@ -701,7 +701,7 @@ def sequana_init(options):
 
     cfg.copy_requirements(target=options.target_dir)
 
-# FIXME If invalid, no error raised
+    # FIXME If invalid, no error raised
     if options.config_params:
         params = [this.strip() for this in options.config_params.split(",")]
         for param in params: