+++ test/scripts/test_sequana_summary.py
@@ -1,36 +1,29 @@
 from sequana.scripts import summary
-from nose.plugins.attrib import attr
 from sequana import sequana_data
 
 
-#@attr("skip")
-class TestPipeline(object):
 
-    @classmethod
-    def setup_class(klass):
-        """This method is run once for each class before any tests are run"""
-        klass.prog = "sequana_summary"
-        klass.params = {'prog': klass.prog}
+prog = "sequana_summary"
 
-    def _test_version(self):
-        summary.main([self.prog, '--version'])
+def _test_version():
+    summary.main([prog, '--version'])
 
-    def test_help(self):
-        try:
-            summary.main([self.prog, '--help'])
-            assert False
-        except SystemExit:
-            pass
-        else:
-            raise Exception
+def test_help():
+    try:
+        summary.main([prog, '--help'])
+        assert False
+    except SystemExit:
+        pass
+    else:
+        raise Exception
 
-    def test_input(self):
-        filename = sequana_data('virus.bed', 'data')
-        df = summary.main([self.prog, '--file', filename])
-        len(df)
+def test_input():
+    filename = sequana_data('virus.bed', 'data')
+    df = summary.main([prog, '--file', filename])
+    len(df)
 
-        filename = sequana_data('test.fastq', "testing")
-        df = summary.main([self.prog, '--file', filename])
+    filename = sequana_data('test.fastq', "testing")
+    df = summary.main([prog, '--file', filename])