+++ test/test_databases.py
@@ -19,14 +19,12 @@ def test_database_download():
         os.remove(this)
     os.rmdir("Viroid")
 
-
     d.download_list()
     # cleanup
     for key in d._metadata.keys():
         res = d._metadata[key]
         filename = res[0]
         os.remove(filename)
-
 
     d.download_accession('AF439431')
     os.remove("Custom/ENA_AF439431.fa")