+++ test/test_snpeff.py
@@ -1,10 +1,8 @@
 from sequana import sequana_data
 from sequana import snpeff
 from easydev import TempFile
-from nose.plugins.attrib import attr
+import os
 
-
-@attr("skip")
 def test_snpeff():
     # a custom refrence
     mydata = snpeff.SnpEff(reference=sequana_data("test_snpeff_ref.gb"))